斷詞拆出去
