+++ tox.ini
@@ -63,6 +63,7 @@ commands=
     python manage.py 教典例句
     coverage run --source=臺灣言語服務 manage.py 訓練Moses正規化 台語
     coverage run --source=臺灣言語服務 -a manage.py test 試驗.翻譯.test台語正規化整合試驗.台語正規化整合試驗
+    coverage run --source=臺灣言語服務 -a manage.py test -p test*斷詞整合試驗.py
 
 [testenv:flake8]
 deps =
